Revert "[spec] Consolidate the per-decode KV alloc reserve into one helper (#27458)"
This reverts commit 1c7acba57937c26c459f07138be2ff4621c4100d.
