+++ python/sglang/srt/debug_utils/pr_fix_toggle.py
@@ -70,15 +70,20 @@ patches:
 
 _PR_REVERT_YAML_26972 = """
 patches:
-  - target: sglang.srt.mem_cache.common.get_req_to_token_extra_context_len
+  - target: sglang.srt.model_executor.model_runner_kv_cache_mixin.ModelRunnerKVCacheMixin._init_pools
     edits:
       - match: |
           if (
-              server_args.speculative_algorithm is not None
-              and server_args.page_size > 1
-              and (server_args.speculative_eagle_topk or 1) > 1
+              self.server_args.speculative_algorithm is not None
+              and self.server_args.page_size > 1
+              and (self.server_args.speculative_eagle_topk or 1) > 1
           ):
-              extra = max(extra, get_alloc_reserve_per_decode(server_args))
+              from sglang.srt.managers.utils import get_alloc_len_per_decode
+
+              extra_max_context_len = max(
+                  extra_max_context_len,
+                  2 * get_alloc_len_per_decode(self.server_args),
+              )
         replacement: ""
 """
